+++ Sources/SuperwallKit/StoreKit/Transactions/Purchasing/ProductPurchaserSK1.swift
@@ -10,34 +10,7 @@ import StoreKit
 
 final class ProductPurchaserSK1: NSObject {
   // MARK: Purchasing
-  actor Purchasing {
-    private var completion: ((PurchaseResult) -> Void)?
-    private var productId: String?
-    var lastTransaction: SKPaymentTransaction?
-
-    func productId(is productId: String) -> Bool {
-      return productId == self.productId
-    }
-
-    func setCompletion(_ completion: @escaping (PurchaseResult) -> Void) {
-      self.completion = completion
-    }
-
-    func beginPurchase(of productId: String) {
-      self.productId = productId
-    }
-
-    func completePurchase(
-      of transaction: SK1Transaction? = nil,
-      result: PurchaseResult
-    ) {
-      lastTransaction = transaction
-      completion?(result)
-      self.completion = nil
-      self.productId = nil
-    }
-  }
-  let purchasing = Purchasing()
+  let coordinator = PurchasingCoordinator()
 
   // MARK: Restoration
   final class Restoration {
@@ -75,12 +48,12 @@ final class ProductPurchaserSK1: NSObject {
   /// Purchases a product, waiting for the completion block to be fired and
   /// returning a purchase result.
   func purchase(product: SKProduct) async -> PurchaseResult {
-    await purchasing.beginPurchase(of: product.productIdentifier)
+    await coordinator.beginPurchase(of: product.productIdentifier)
 
     let task = Task {
       return await withCheckedContinuation { continuation in
         Task {
-          await purchasing.setCompletion { result in
+          await coordinator.setCompletion { result in
             continuation.resume(returning: result)
           }
         }
@@ -144,10 +117,10 @@ extension ProductPurchaserSK1: SKPaymentTransactionObserver {
       let isPaywallPresented = Superwall.shared.isPaywallPresented
       let paywallViewController = Superwall.shared.paywallViewController
       for transaction in transactions {
+        await coordinator.storeIfPurchased(transaction)
         await checkForTimeout(of: transaction, in: paywallViewController)
         await updatePurchaseCompletionBlock(for: transaction)
         await checkForRestoration(transaction, isPaywallPresented: isPaywallPresented)
-
         Task(priority: .background) {
           await record(transaction)
         }
@@ -209,13 +182,17 @@ extension ProductPurchaserSK1: SKPaymentTransactionObserver {
           withProductId: skTransaction.payment.productIdentifier
         )
         SKPaymentQueue.default().finishTransaction(skTransaction)
-        await purchasing.completePurchase(
+
+        await coordinator.completePurchase(
           of: skTransaction,
           result: .purchased
         )
       } catch {
         SKPaymentQueue.default().finishTransaction(skTransaction)
-        await purchasing.completePurchase(result: .failed(error))
+        await coordinator.completePurchase(
+          of: skTransaction,
+          result: .failed(error)
+        )
       }
     case .failed:
       SKPaymentQueue.default().finishTransaction(skTransaction)
@@ -224,7 +201,10 @@ extension ProductPurchaserSK1: SKPaymentTransactionObserver {
           switch error.code {
           case .paymentCancelled,
             .overlayCancelled:
-            return await purchasing.completePurchase(result: .cancelled)
+            return await coordinator.completePurchase(
+              of: skTransaction,
+              result: .cancelled
+            )
           default:
             break
           }
@@ -232,17 +212,22 @@ extension ProductPurchaserSK1: SKPaymentTransactionObserver {
           if #available(iOS 14, *) {
             switch error.code {
             case .overlayTimeout:
-              await purchasing.completePurchase(result: .cancelled)
+              await coordinator.completePurchase(
+                of: skTransaction,
+                result: .cancelled
+              )
             default:
               break
             }
           }
         }
-        await purchasing.completePurchase(result: .failed(error))
+        await coordinator.completePurchase(
+          of: skTransaction,
+          result: .failed(error))
       }
     case .deferred:
       SKPaymentQueue.default().finishTransaction(skTransaction)
-      await purchasing.completePurchase(of: skTransaction, result: .pending)
+      await coordinator.completePurchase(of: skTransaction, result: .pending)
     default:
       break
     }